% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grad.logistic.R
\name{grad.logistic}
\alias{grad.logistic}
\title{Gradient of the logistic regression risk.}
\usage{
grad.logistic(x, opts)
}
\arguments{
\item{x}{A p-dimensional vector where the gradient is computed.}

\item{opts}{List of parameters, which must include: \itemize{ \item \code{A},
  a n*p design matrix \item \code{b}, a n-dimensional response vector taking
  values \eqn{1} or \eqn{-1}. }}
}
\value{
The gradient of the function \eqn{f(x) = 1/n sum_{i=1}^n ( log(1 +
  exp(- b_i * x' * A_i) )}, which is: \deqn{f'(x) = 1/n sum_{i=1}^n -b_i A_i
  / (1 + exp( b_i * x' * A_i ))}
}
\description{
Computes the gradient of the logistic regression error: \deqn{f(x) = 1/n
sum_{i=1}^n ( log(1 + exp(- b_i * x' * A_i) ) ,} for a given design matrix
\code{A} and a response binary vector \code{b} with values \code{+1} and
\code{-1}.
}
\examples{
grad.logistic(c(1,3,-2), list(A=diag(3), b=c(1,-1,1))
# The following give the same (elastic net penalized logistic regression without offset):
glmnet(A,b,family="binomial",lambda=0.1,standardize=FALSE,intercept=FALSE, alpha=0.5)$b
apg(grad.logistic, prox.elasticnet, ncol(A), list(A=A, b=b, lambda=0.1, alpha=0.5) )$x
}

