% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad.rankinglogistic}
\alias{grad.rankinglogistic}
\title{Gradient of the ranking logistic regression risk.}
\usage{
grad.rankinglogistic(x, opts)
}
\arguments{
\item{x}{A p-dimensional vector where the gradient is computed.}

\item{opts}{List of parameters, which must include: \itemize{ \item \code{A},
a n*p design matrix \item \code{comp}, a list of length n, where the i-th entry is a vector containing the indices of the samples j which are larger than sample i.}}
}
\value{
The gradient of the function \eqn{f(x) = 1/|comp|
sum_{(i,j)\in comp} ( log(1 + exp(- x' * (A_j - A_i) )}, which is: \deqn{f'(x) = 1/|comp| sum_{(i,j)\in comp} (A_j - A_i) / (1 + exp( x' * ( A_i - A_j) )}
}
\description{
Computes the gradient of the ranking logistic regression error: \deqn{f(x) = 1/|comp|
sum_{(i,j)\in comp} ( log(1 + exp(- x' * (A_j - A_i) )} for a given design matrix
\code{A} and a set of pairs \eqn{(i,j)\in P} where the j-th sample is larger than the i-th sample.
}
\examples{
grad.rankinglogistic(c(1,3,-2), list(A=diag(3), comp=list(as.integer(c(2,3)),as.integer(3),integer(length=0))))
}

