% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_vec.R
\name{norm_vec}
\alias{norm_vec}
\title{Euclidean norm of a vector}
\usage{
norm_vec(x)
}
\arguments{
\item{x}{The input vector}
}
\value{
The Euclidean norm of the input vector
}
\description{
Computes the Euclidean norm of a vector
}
\examples{
norm_vec(c(1,1))
}

