% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prox.boundednondecreasing.R
\name{prox.boundednondecreasing}
\alias{prox.boundednondecreasing}
\title{Proximal operator for the set of bounded non-decreasing vectors}
\usage{
prox.boundednondecreasing(x, t = 0, opts = list())
}
\arguments{
\item{x}{The input vector}

\item{opts}{Optional list of parameters, which can include: \itemize{ \item
  \code{maxnorm} : the bound on the Euclidean norm of the non-decreasing
  vector (default \code{1}).}}
}
\value{
The projection of \code{x} onto the set of nondecreasing vectors with
  norm bounded by \code{maxnorm}.
}
\description{
Computes the proximal operator of the bounded non-decreasing vectors, i.e.,
\deqn{prox(x) = argmin_u || u - x ||^2 such that u_1 \le ... \le u_n and
||u|| \le M} This is simply the projection of $x$ onto the set of
non-decreasing vectors, obtained by isotonic regression, followed by
projection onto the Euclidean ball of radius $M$.
}
\examples{
a <- rnorm(20)+1
plot(a)
lines(prox.boundednondecreasing(a), col=2)
lines(prox.boundednondecreasing(a, opts=list(maxnorm=10)), col=3)
}

