% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.elasticnet.R
\name{prox.elasticnet}
\alias{prox.elasticnet}
\title{Proximal operator of the scaled elastic net penalty.}
\usage{
prox.elasticnet(x, t = 1, opts = list(lambda = 1, alpha = 1))
}
\arguments{
\item{x}{The input vector}

\item{t}{The step size (default is \code{1})}

\item{opts}{List of parameters, which can include: \itemize{ \item
\code{lambda} : the scaling factor of the L1 norm (default is
\code{lambda=1})\item \code{alpha} : the balance between L1 and L2 norms.
\code{alpha=0} is the squared L2 (ridge) penalty, \code{alpha=1} is the L1
(lasso) penalty. Default is \code{alpha=1} (lasso).}}
}
\value{
The proximal of \eqn{h} at {x} with step size \eqn{t}, given by
  \deqn{prox_h(x,t) = argmin_u [ t h(u) + 1/2 || x - u ||^2 ]}.
}
\description{
Computes the proximal operator of the scaled elastic net penalty: \deqn{h(x)
= \lambda [ (1 - \alpha)/2 ||x||_2^2 + \alpha ||x||_1 ] ,} where
\eqn{\lambda} is a scaling factor and \eqn{\alpha \in [0,1]} balances between
the L1 and L2 norms.
}
\examples{
prox.elasticnet(c(1,3,-2), 1.5, list(lambda=1,alpha=0.5))
}

