% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.grouplasso.R
\name{prox.grouplasso}
\alias{prox.grouplasso}
\title{Proximal operator of the group lasso penalty}
\usage{
prox.grouplasso(x, t, opts = list(groups = as.list(seq(length(xc)))))
}
\arguments{
\item{x}{The input vector}

\item{t}{The step size}

\item{opts}{List of parameters, which can include: \itemize{ \item
\code{groups} : a list of groups, each group is just a sequence of indices
of the components that form the group (default: all singletons). \item \code{groupweigths} : a vector of weights for the groups. If a single number, all groups have the same weight (default \code{1})
}}
}
\value{
The proximal operator of the group lasso, which is a soft-thresholing
  operator applied to the restriction of the \code{x} to each group.
}
\description{
Computes the proximal operator of the group lasso penalty: \deqn{h(x) =
\sum_{g group} w_g ||x[g]||_2 .} Note that the groups should not
overlap.
}
\examples{
x <- rnorm(5)
# When groups are all the singletons we recover the L1 (lasso) penalty
prox.grouplasso(f,1,list(groups=as.list(seq(length(f)))))
prox.elasticnet(f,1,list(lambda=1,alpha=1))
}

