% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prox.isotonic.R
\name{prox.isotonic}
\alias{prox.isotonic}
\title{Proximal operator of the isotonic constraint}
\usage{
prox.isotonic(x, ...)
}
\arguments{
\item{x}{The input vector}
}
\value{
The projection of \code{x} onto the set of nondecreasing vectors,
  obtained by solving an isotonic regression.
}
\description{
Computes the proximal operator of the isotonic constraint, i.e., the
projection onto the set of nondecreasing vectors.
}
\examples{
prox.isotonic(c(1,3,-2,4,5))
}

