% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prox.l1.R
\name{prox.l1}
\alias{prox.l1}
\title{Proximal operator of the scaled L1 norm.}
\usage{
prox.l1(x, t, opts)
}
\arguments{
\item{x}{The input vector}

\item{t}{The step size}

\item{opts}{List of parameters, which must include:
  \itemize{\code{opts$lambda}, the scaling factor of the L1 norm.}}
}
\value{
The proximal of \eqn{h} at {x} with step size \eqn{t}, given by
  \deqn{prox_h(x,t) = argmin_u [ t h(u) + 1/2 || x - u ||^2 ]}.
}
\description{
Computes the proximal operator of the L1 norm: \deqn{h(x) = \lambda ||x||_1
,} where \eqn{\lambda} is a scaling factor.
}
\examples{
prox.l1(c(1,3,-2), 1.5, list(lambda=1))
}

