% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_to_pairs.R
\name{surv_to_pairs}
\alias{surv_to_pairs}
\title{Given survival information (time and status) for a set of subjects, compute all pairs (i,j) for which patient i has the event before patient j.}
\usage{
surv_to_pairs(time, event = rep(1, length(time)))
}
\arguments{
\item{time}{A vector of follow up time for the subjects}

\item{event}{A vector of status indicator for the subjects, normally \code{0}=alive, \code{1}=dead. Other choices are \code{TRUE/FALSE} (\code{TRUE}=death). By default, all subjects are assumed to have an event.}
}
\value{
A list of vectors, where the i-th vector is the set of patients that have the events after patient i.
}
\description{
Given survival information (time and status) for a set of subjects, compute all pairs (i,j) for which patient i has the event before patient j.
}
\examples{
time <- c(5,6,7,7,8,9)
event <- c(1,0,0,1,1,0)
res <- surv_to_pairs(time,event)
}

